things_rest/requirements.txt
Django~=3.0.7
django-rest-knox
djangorestframework